+++ src/main/java/com/api/comic_reader/repositories/RatingRepository.java
@@ -20,7 +20,7 @@ public interface RatingRepository extends JpaRepository<RatingEntity, Long> {
             @Param("comic") ComicEntity comic, @Param("user") UserEntity user, @Param("score") Long score);
 
     @Query("select avg(r.score) from RatingEntity r where r.comic = :comic")
-    Long getComicAverageRating(@Param("comic") ComicEntity comic);
+    Double getComicAverageRating(@Param("comic") ComicEntity comic);
 
     @Query("select r from RatingEntity r where r.comic = :comic and r.user = :user")
     RatingEntity findByComicAndUser(@Param("comic") ComicEntity comic, @Param("user") UserEntity user);
